Turn alignment options into enums
