+++ examples/specials.rs
@@ -11,7 +11,7 @@ use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, Window,
 };
 use embedded_text::{
-    alignment::Justified,
+    alignment::HorizontalAlignment,
     style::{HeightMode, TextBoxStyleBuilder},
     TextBox,
 };
@@ -38,7 +38,7 @@ fn main() {
         .build();
 
     let textbox_style = TextBoxStyleBuilder::new()
-        .alignment(Justified)
+        .alignment(HorizontalAlignment::Justified)
         .height_mode(HeightMode::FitToText)
         .build();
