+++ src/alignment/mod.rs
@@ -1,60 +1,156 @@
 //! Text alignment options.
 use crate::{
+    parser::SPEC_CHAR_NBSP,
     rendering::{cursor::Cursor, space_config::SpaceConfig},
     style::LineMeasurement,
+    utils::str_width,
     TextBox,
 };
-use embedded_graphics::text::renderer::TextRenderer;
+use embedded_graphics::{prelude::Dimensions, text::renderer::TextRenderer};
 
-pub mod bottom;
-pub mod center;
-pub mod justified;
-pub mod left;
-pub mod right;
-pub mod scrolling;
-pub mod top;
+#[cfg(test)]
+mod test;
 
-/// Horizontal text alignment base trait.
-///
-/// Use implementors to parametrize [`TextBoxStyle`] and [`TextBoxStyleBuilder`].
-///
-/// [`TextBoxStyle`]: ../style/struct.TextBoxStyle.html
-/// [`TextBoxStyleBuilder`]: ../style/builder/struct.TextBoxStyleBuilder.html
-pub trait HorizontalTextAlignment: Copy {
-    /// Type of the associated whitespace information.
-    type SpaceConfig: SpaceConfig;
+/// Horizontal text alignment options.
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
+pub enum HorizontalAlignment {
+    /// Left aligned.
+    Left,
 
-    /// Whether to move cursor via leading spaces in the first line.
-    const IGNORE_LEADING_SPACES: bool = true;
+    /// Center aligned.
+    Center,
+
+    /// Right aligned.
+    Right,
+
+    /// Fully justified.
+    Justified,
+}
+
+impl HorizontalAlignment {
+    pub(crate) fn ignores_leading_spaces(self) -> bool {
+        match self {
+            HorizontalAlignment::Left => false,
+            HorizontalAlignment::Center => true,
+            HorizontalAlignment::Right => true,
+            HorizontalAlignment::Justified => true,
+        }
+    }
 
     /// Calculate offset from the left side and whitespace information.
-    fn place_line(
+    pub fn place_line(
+        self,
         line: &str,
         renderer: &impl TextRenderer,
         measurement: LineMeasurement,
-    ) -> (u32, Self::SpaceConfig);
+    ) -> (u32, SpaceConfig) {
+        match self {
+            HorizontalAlignment::Left => (0, SpaceConfig::new_from_renderer(renderer)),
+            HorizontalAlignment::Center => (
+                (measurement.max_line_width - measurement.width + 1) / 2,
+                SpaceConfig::new_from_renderer(renderer),
+            ),
+            HorizontalAlignment::Right => (
+                measurement.max_line_width - measurement.width,
+                SpaceConfig::new_from_renderer(renderer),
+            ),
+            HorizontalAlignment::Justified => {
+                let space_width = str_width(renderer, " ");
+                let space_chars = [' ', SPEC_CHAR_NBSP];
+
+                let mut space_count = 0;
+                let mut partial_space_count = 0;
+
+                for c in line.chars().skip_while(|c| space_chars.contains(c)) {
+                    if space_chars.contains(&c) {
+                        partial_space_count += 1;
+                    } else {
+                        space_count += partial_space_count;
+                        partial_space_count = 0;
+                    }
+                }
+
+                let space_info = if !measurement.last_line && space_count != 0 {
+                    let space =
+                        measurement.max_line_width - measurement.width + space_count * space_width;
+                    let space_width = space / space_count;
+                    let extra_pixels = space % space_count;
+                    SpaceConfig::new(space_width, Some(extra_pixels))
+                } else {
+                    SpaceConfig::new(space_width, None)
+                };
+                (0, space_info)
+            }
+        }
+    }
 }
 
-/// Vertical text alignment base trait.
-///
-/// Use implementors to parametrize [`TextBoxStyle`] and [`TextBoxStyleBuilder`].
-///
-/// [`TextBoxStyle`]: ../style/struct.TextBoxStyle.html
-/// [`TextBoxStyleBuilder`]: ../style/builder/struct.TextBoxStyleBuilder.html
-pub trait VerticalTextAlignment: Copy {
+/// Vertical text alignment options.
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
+pub enum VerticalAlignment {
+    /// Top aligned text.
+    ///
+    /// The first line of the text will be rendered at the top of the text box.
+    Top,
+
+    /// Middle aligned text.
+    ///
+    /// The text will be vertically centered within the text box.
+    Middle,
+
+    /// Bottom aligned text.
+    ///
+    /// The last line of the text will be aligned to the bottom of the text box.
+    Bottom,
+
+    /// Scrolling alignment.
+    ///
+    /// Aligns the last line of the text to be always visible. If the text fits inside the text box,
+    /// it will be top aligned. If the text is longer, it will be bottom aligned.
+    Scrolling,
+}
+
+impl VerticalAlignment {
     /// Set the cursor's initial vertical position
-    fn apply_vertical_alignment<'a, S, A>(
+    pub fn apply_vertical_alignment<'a, S>(
+        self,
         cursor: &mut Cursor,
-        styled_text_box: &'a TextBox<'a, S, A, Self>,
+        styled_text_box: &'a TextBox<'a, S>,
     ) where
         S: TextRenderer,
-        A: HorizontalTextAlignment;
-}
+    {
+        let text_height = styled_text_box.style.measure_text_height(
+            &styled_text_box.character_style,
+            styled_text_box.text,
+            cursor.line_width(),
+        ) as i32;
 
-pub use bottom::BottomAligned;
-pub use center::CenterAligned;
-pub use justified::Justified;
-pub use left::LeftAligned;
-pub use right::RightAligned;
-pub use scrolling::Scrolling;
-pub use top::TopAligned;
+        let box_height = styled_text_box.bounding_box().size.height as i32;
+
+        match self {
+            VerticalAlignment::Top => {
+                // nothing to do here
+            }
+
+            VerticalAlignment::Middle => {
+                let offset = (box_height - text_height) / 2;
+
+                cursor.y += offset;
+            }
+
+            VerticalAlignment::Bottom => {
+                let offset = box_height - text_height;
+
+                cursor.y += offset;
+            }
+
+            VerticalAlignment::Scrolling => {
+                if text_height > box_height {
+                    let offset = box_height - text_height;
+
+                    cursor.y += offset
+                }
+            }
+        }
+    }
+}
