+++ src/alignment/test/mod.rs
@@ -1,0 +1,6 @@
+mod center;
+mod justified;
+mod left;
+mod right;
+mod scrolling;
+mod top;
