+++ src/style/builder.rs
@@ -2,7 +2,7 @@
 use embedded_graphics::text::LineHeight;
 
 use crate::{
-    alignment::{HorizontalTextAlignment, LeftAligned, TopAligned, VerticalTextAlignment},
+    alignment::{HorizontalAlignment, VerticalAlignment},
     style::{HeightMode, TabSize, TextBoxStyle, VerticalOverdraw},
 };
 
@@ -10,35 +10,33 @@ use crate::{
 ///
 /// [`TextBoxStyle`]: struct.TextBoxStyle.html
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
-pub struct TextBoxStyleBuilder<A, V> {
-    style: TextBoxStyle<A, V>,
+pub struct TextBoxStyleBuilder {
+    style: TextBoxStyle,
 }
 
-impl Default for TextBoxStyleBuilder<LeftAligned, TopAligned> {
+impl Default for TextBoxStyleBuilder {
     #[inline]
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl TextBoxStyleBuilder<LeftAligned, TopAligned> {
+impl TextBoxStyleBuilder {
     /// Creates a new text box style builder object.
     #[inline]
     #[must_use]
     pub const fn new() -> Self {
         Self {
             style: TextBoxStyle {
-                alignment: LeftAligned,
-                vertical_alignment: TopAligned,
+                alignment: HorizontalAlignment::Left,
+                vertical_alignment: VerticalAlignment::Top,
                 height_mode: HeightMode::Exact(VerticalOverdraw::FullRowsOnly),
                 line_height: LineHeight::Percent(100),
                 tab_size: TabSize::Spaces(4),
             },
         }
     }
-}
 
-impl<A, V> TextBoxStyleBuilder<A, V> {
     /// Sets the line height.
     ///
     /// The line height is defined as the vertical distance between the baseline of two adjacent lines
@@ -56,7 +54,7 @@ impl<A, V> TextBoxStyleBuilder<A, V> {
     /// ```
     #[inline]
     #[must_use]
-    pub fn line_height(mut self, line_height: LineHeight) -> Self {
+    pub const fn line_height(mut self, line_height: LineHeight) -> Self {
         self.style.line_height = line_height;
 
         self
@@ -65,43 +63,28 @@ impl<A, V> TextBoxStyleBuilder<A, V> {
     /// Sets the horizontal text alignment.
     #[inline]
     #[must_use]
-    pub fn alignment<TA: HorizontalTextAlignment>(
-        self,
-        alignment: TA,
-    ) -> TextBoxStyleBuilder<TA, V> {
-        TextBoxStyleBuilder {
-            style: TextBoxStyle {
-                alignment,
-                line_height: self.style.line_height,
-                vertical_alignment: self.style.vertical_alignment,
-                height_mode: self.style.height_mode,
-                tab_size: self.style.tab_size,
-            },
-        }
+    pub const fn alignment(mut self, alignment: HorizontalAlignment) -> TextBoxStyleBuilder {
+        self.style.alignment = alignment;
+
+        self
     }
 
     /// Sets the vertical text alignment.
     #[inline]
     #[must_use]
-    pub fn vertical_alignment<VA: VerticalTextAlignment>(
-        self,
-        vertical_alignment: VA,
-    ) -> TextBoxStyleBuilder<A, VA> {
-        TextBoxStyleBuilder {
-            style: TextBoxStyle {
-                alignment: self.style.alignment,
-                line_height: self.style.line_height,
-                vertical_alignment,
-                height_mode: self.style.height_mode,
-                tab_size: self.style.tab_size,
-            },
-        }
+    pub const fn vertical_alignment(
+        mut self,
+        vertical_alignment: VerticalAlignment,
+    ) -> TextBoxStyleBuilder {
+        self.style.vertical_alignment = vertical_alignment;
+
+        self
     }
 
     /// Sets the height mode.
     #[inline]
     #[must_use]
-    pub fn height_mode(mut self, height_mode: HeightMode) -> TextBoxStyleBuilder<A, V> {
+    pub const fn height_mode(mut self, height_mode: HeightMode) -> TextBoxStyleBuilder {
         self.style.height_mode = height_mode;
         self
     }
@@ -109,35 +92,25 @@ impl<A, V> TextBoxStyleBuilder<A, V> {
     /// Sets the tab size.
     #[inline]
     #[must_use]
-    pub fn tab_size(mut self, tab_size: TabSize) -> Self {
+    pub const fn tab_size(mut self, tab_size: TabSize) -> Self {
         self.style.tab_size = tab_size;
 
         self
     }
-}
 
-impl<A, V> TextBoxStyleBuilder<A, V>
-where
-    A: HorizontalTextAlignment,
-    V: VerticalTextAlignment,
-{
     /// Builds the [`TextBoxStyle`].
     ///
     /// [`TextBoxStyle`]: struct.TextBoxStyle.html
     #[inline]
     #[must_use]
-    pub fn build(self) -> TextBoxStyle<A, V> {
+    pub const fn build(self) -> TextBoxStyle {
         self.style
     }
 }
 
-impl<A, V> From<&TextBoxStyle<A, V>> for TextBoxStyleBuilder<A, V>
-where
-    A: Copy,
-    V: Copy,
-{
+impl From<&TextBoxStyle> for TextBoxStyleBuilder {
     #[inline]
-    fn from(style: &TextBoxStyle<A, V>) -> Self {
+    fn from(style: &TextBoxStyle) -> Self {
         Self { style: *style }
     }
 }
